DataProcessor/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Scraping
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
newspaper3k>=0.2.8
trafilatura>=1.5.0

# Google Search
google-api-python-client>=2.95.0

# LLM Connection
requests>=2.31.0
websockets>=11.0.3

# RSS Feed
feedparser==6.0.10
bs4>=0.0.1
python-dateutil==2.8.2

# Utils
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
